Updated folder structure and refactored the code
