+++ src/templates/vue/src/bootstrap/components/atoms/Label.vue
