+++ src/templates/vue/src/bootstrap/components/atoms/LogoutButton.vue
