+++ src/templates/vue/src/bootstrap/components/organisms/Navbar.vue
