+++ src/templates/vue/src/bootstrap/components/organisms/Table.vue
