+++ src/templates/vue/src/bootstrap/pages/Home.vue
