--- src/templates/vue/src/components/tailwind/pages/Home.vue
@@ -1,13 +1,0 @@
-<template>
-  <div>
-    <h1 class="text-4xl dark:text-white dark:bg-vueBlack">
-      Welcome to Home Screen
-    </h1>
-  </div>
-</template>
-
-<script>
-export default {
-  name: "Home",
-};
-</script>
