+++ src/templates/vue/src/tailwind/components/atoms/Input.vue
