+++ src/templates/vue/src/tailwind/components/atoms/Label.vue
