+++ src/templates/vue/src/tailwind/components/atoms/LogoutButton.vue
