+++ src/templates/vue/src/tailwind/components/organisms/Table.vue
