+++ src/templates/vue/src/tailwind/pages/Home.vue
@@ -1,0 +1,11 @@
+<template>
+  <h1 class="text-4xl dark:text-white dark:bg-vueBlack">
+    Welcome to Home Screen
+  </h1>
+</template>
+
+<script>
+export default {
+  name: "Home",
+};
+</script>
