+++ src/templates/vue/src/tailwind/styles/index.css
