+++ src/vuexTemplates/tailwind/userModal/EditUser.vue
@@ -19,7 +19,8 @@
             id="name-input"
             type="text"
             placeholder="Name"
-            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded border border-slate-300 outline-none focus:outline-none focus:ring w-full"
+            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded border 
+            border-slate-300 outline-none focus:outline-none focus:ring w-full"
             v-model="name"
           />
         </div>
@@ -29,7 +30,8 @@
             id="username-input"
             type="text"
             placeholder="Username"
-            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded border border-slate-300 outline-none focus:outline-none focus:ring w-full"
+            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded border 
+            border-slate-300 outline-none focus:outline-none focus:ring w-full"
             v-model="username"
           />
         </div>
@@ -38,7 +40,8 @@
           <input
             id="email-input"
             type="email"
-            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded border border-slate-300 outline-none focus:outline-none focus:ring w-full"
+            class="px-3 py-5 mt-1 placeholder-slate-300 text-slate-600 relative h-10 rounded 
+            border border-slate-300 outline-none focus:outline-none focus:ring w-full"
             placeholder="example@email.com"
             v-model="email"
           />
@@ -56,7 +59,8 @@
         name="Submit"
       />
       <Button
-        class="text-gray border border-gray-500 px-3 py-2 ml-auto m-3 shadow-sm hover:bg-gray-500 hover:text-white active:bg-gray-600 font-bold rounded outline-none focus:outline-none dark:text-white"
+        class="text-gray border border-gray-500 px-3 py-2 ml-auto m-3 shadow-sm hover:bg-gray-500 
+        hover:text-white active:bg-gray-600 font-bold rounded outline-none focus:outline-none dark:text-white"
         type="button"
         name="Cancel"
         @onClick="close"
